added register payload

+++ client/message_u_client/message_u_client/Payload.h
@@ -8,5 +8,5 @@ public:
 	using bytes_t = std::vector<uint8_t>;
 
 	virtual bytes_t toBytes() = 0;
-	virtual size_t getSize() = 0;
+	virtual uint32_t getSize() = 0;
 };
